+++ .travis.yml
@@ -2,8 +2,12 @@ language: go
 
 go:
   - 1.x
-  
+
 sudo: false
 
 install:
   - go get -t -v ./...
+
+after_success:
+  - npm install netlify-cli -g
+  - netlify deploy --dir=tmpl
